+++ Changelog.md
@@ -30,6 +30,7 @@ If so, please delete it since it will prevent the federation from working proper
 * Fix height too high on mobile SPV [#7480](https://github.com/diaspora/diaspora/pull/7480)
 * Improve stream when ignoring a person who posts a lot of tagged posts [#7503](https://github.com/diaspora/diaspora/pull/7503)
 * Fix order of comments across pods [#7436](https://github.com/diaspora/diaspora/pull/7436)
+* Prevent publisher from closing in preview mode [#7518](https://github.com/diaspora/diaspora/pull/7518)
 
 ## Features
 * Add support for mentions in comments to the backend [#6818](https://github.com/diaspora/diaspora/pull/6818)